requirements.txt
Flask==3.0.0
google-cloud-storage==2.18.2
google-cloud-datastore==2.15.1
python-jose
six
requests
authlib